com.unity.memoryprofiler@0.2.6-preview.1 ## [0.2.6-preview.1] - 2020-10-01 ### Added   - Added search field delay when filtering table entries via the match filter.   - Added a dialog box when attempting to rename a snapshot to same name as an existing one ([case 1273417](https://fogbugz.unity3d.com/f/cases/1273417/)).   - Added a dialog box when attempting to rename a snapshot with invalid characters ([case 1274987](https://fogbugz.unity3d.com/f/cases/1274987/)).   - Added name tags for snapshot preview screenshots, to avoid confusion when the user takes a snapshot with the memory profiler window open.   - Added name tag for memory map backing memory texture.
### Changed
  - Removed unneeded finalizer from CachedSnasphot, which could crash the Editor.
  - Fixed an issue inside the VM data validation tools, where we would not catch invalid VM info data.
  - Fixed incorrect parsing of table headers as callstack site ids when attempting to retrieve callstacks for a native allocation.
  - Fixed texture color space to linear inside the MemoryMap view creation ([case 1261948](https://fogbugz.unity3d.com/f/cases/1261948/)).
  - Fixed potential alpha overflows inside the MemoryMap, where some regions on the memory map might have become transparent.
  - Secured the MemorySnapshot backend so that we handle array sizes larger than int.Max gracefully.
  - Fixed an issue with reading entries null entries from reference arrays, where this would cause a NullReferenceException.
  - Changed the snapshot crawler to discard unknown type data as some types are VM internal.
    Note: The following versions are able to capture all type data: 2020.2b2, 2020.1.5f1, 2019.4.10f1.
  - Changed table diffing to be multi-key, thus making the diffing operation accurate with regards to matching object as "Same".
  - Changed the Address column formatting to show pointer values in hex rather than decimal.
  - Changed the Name column so that it only shows object or field name depending on the expanded item.
  - Fixed some static fields not showing up in the Reference table.
  - Fixed a leak where the snapshot collection would never clean up it's Texture2D.
  - Fixed a leak where the memory map's backing texture would never be cleaned up during flush.
  - Fixed an issue where snapshots could be imported twice ([case 1271886](https://fogbugz.unity3d.com/f/cases/1271886/)).
  - Fixed an issue showing "None" in the diff table's match filter, when matching by "Diff" column value.
  - Fixed an issue where the snapshot deletion message would not to explicitly state that the file will be permanently removed.
  - Fixed an issue where the naming convention used for snapshots inside the memory map diff view was incorrect.
  - Fixed an issue where the snapshots used inside memory map would not swap when the swap button was pressed.
  - Fixed an issue where snapshot file list was sorted alphabetically ([case 1276092](https://fogbugz.unity3d.com/f/cases/1276092/)).
  - Fixed an issue where open snapshots could end up losing their open state while still open in the tool ([case 1275288](https://fogbugz.unity3d.com/f/cases/1275288/)).
  - Fixed an issue where the snapshot file list wouldn't update when the editor would change focus ([case 1275835](https://fogbugz.unity3d.com/f/cases/1275835/)).

+++ Editor/MemorySnapshot/BlockSection.cs
@@ -40,7 +40,7 @@ namespace Unity.MemoryProfiler.Editor.Format
         // effectively bypassing the need for getItemFunc and the temp byte array.
         // However, Stream.CopyTo(Stream, int) is not available as of .NET runtime 3.5, making our only
         // option the copy into the temp byte array.
-        internal void GetData(ulong startBlockOffset, uint blockLength, ref byte[] dataOut, BinaryReader reader)
+        internal void GetData(ulong startBlockOffset, ulong blockLength, ref byte[] dataOut, BinaryReader reader)
         {
             Debug.Assert(startBlockOffset + blockLength <= m_TotalBytes);
